+++ examples/general_examples.jl
@@ -30,7 +30,7 @@ function generate_perpendicular_lines(steps::Int, minlen, maxlen)
     rec([0 0], [0 1], steps)
     vnum = size(V, 1)
     enum = vnum >> 1
-    EV = spzeros(Int8, enum, vnum)
+    EV = spzeros(Int, enum, vnum)
     for i in 1:enum
         EV[i, i*2-1:i*2] = 1
     end
@@ -47,7 +47,7 @@ function generate_random_lines(n, points_range, alphas_range)
         new_points[i, :] = origins[i, :] + alphas[i]*directions[i, :]
     end
     V = [origins; new_points]
-    EV = spzeros(Int8, n, n*2)
+    EV = spzeros(Int, n, n*2)
     for i in 1:n
         EV[i, i] = 1
         EV[i, n+i] = 1
@@ -63,7 +63,7 @@ function rubiks_example(ncubes = 3)
         1 1 1; 1 0 1
     ]
 
-    EV = sparse(Int8[
+    EV = sparse(Int[
         -1  1  0  0  0  0  0  0;
         0 -1  1  0  0  0  0  0;
         0  0 -1  1  0  0  0  0;
@@ -78,7 +78,7 @@ function rubiks_example(ncubes = 3)
         0  0  0  0 -1  0  0  1;
     ])
 
-    FE = sparse(Int8[
+    FE = sparse(Int[
         1  1  1 -1  0  0  0  0  0  0  0  0;
         0  0  0  0  0  0  0  0 -1 -1 -1  1;
         -1  0  0  0  1 -1  0  0  1  0  0  0;
@@ -88,7 +88,7 @@ function rubiks_example(ncubes = 3)
     ])
 
     cube = [LinearAlgebraicRepresentation.Points(V), EV, FE]
-    cubesRow = [LinearAlgebraicRepresentation.Points(zeros(0,3)),spzeros(Int8,0,0),spzeros(Int8,0,0)]
+    cubesRow = [LinearAlgebraicRepresentation.Points(zeros(0,3)),spzeros(Int,0,0),spzeros(Int,0,0)]
 
     println(map(typeof, cube))
     println(map(typeof, cubesRow))
